+++ src/model/bar.ts
@@ -1,5 +1,5 @@
 import type { MaidrLayer } from '@type/maidr';
-import type { AudioState, TextState } from '@type/state';
+import type { AudioState, HighlightState, TextState } from '@type/state';
 import type { BarPoint } from './grammar';
 import { Orientation } from '@type/plot';
 import { AbstractTrace } from './plot';
@@ -7,8 +7,11 @@ import { AbstractTrace } from './plot';
 export abstract class AbstractBarPlot<T extends BarPoint> extends AbstractTrace<number> {
   protected readonly points: T[][];
   protected readonly barValues: number[][];
-  protected readonly orientation: Orientation;
 
+  protected readonly brailleValues: string[][];
+  protected readonly highlightValues: SVGElement[][];
+
+  protected readonly orientation: Orientation;
   protected readonly min: number[];
   protected readonly max: number[];
 
@@ -27,7 +30,21 @@ export abstract class AbstractBarPlot<T extends BarPoint> extends AbstractTrace<
     this.min = this.barValues.map(row => Math.min(...row));
     this.max = this.barValues.map(row => Math.max(...row));
 
-    this.brailleValues = this.getBraille();
+    this.brailleValues = this.mapToBraille(this.barValues);
+    this.highlightValues = this.mapToSvgElements(layer.selectors as string);
+  }
+
+  public dispose(): void {
+    this.points.length = 0;
+    this.barValues.length = 0;
+
+    this.brailleValues.length = 0;
+    this.highlightValues.length = 0;
+
+    this.min.length = 0;
+    this.max.length = 0;
+
+    super.dispose();
   }
 
   protected get values(): number[][] {
@@ -67,8 +84,23 @@ export abstract class AbstractBarPlot<T extends BarPoint> extends AbstractTrace<
     };
   }
 
-  protected getBraille(): string[][] {
-    return this.barValues.map((row, index) =>
+  protected highlight(): HighlightState {
+    if (this.highlightValues.length === 0) {
+      return {
+        empty: true,
+        type: 'trace',
+        traceType: this.type,
+      };
+    }
+
+    return {
+      empty: false,
+      elements: this.highlightValues[this.row][this.col],
+    };
+  }
+
+  private mapToBraille(data: number[][]): string[][] {
+    return data.map((row, index) =>
       this.createBraille(row, this.min[index], this.max[index]),
     );
   }
@@ -96,6 +128,24 @@ export abstract class AbstractBarPlot<T extends BarPoint> extends AbstractTrace<
     }
 
     return braille;
+  }
+
+  protected mapToSvgElements(selector?: string): SVGElement[][] {
+    if (!selector) {
+      return new Array<Array<SVGElement>>();
+    }
+
+    const svgElements = [Array.from(document.querySelectorAll<SVGElement>(selector))];
+    if (svgElements.length !== this.points.length) {
+      return new Array<Array<SVGElement>>();
+    }
+    for (let row = 0; row < this.points.length; row++) {
+      if (svgElements[row].length !== this.points[row].length) {
+        return new Array<Array<SVGElement>>();
+      }
+    }
+
+    return svgElements;
   }
 }