+++ src/type/color.ts
@@ -1,0 +1,5 @@
+export interface Rgb {
+  r: number;
+  g: number;
+  b: number;
+}